Decouple GUI from Business Logic
Refactor code in order to decouple the business logic from the
presentation layer. The Job class is now the base class for the business
logic of the program, and two subclasses were created, one to be used
with the GUI app and the other with the CLI app.
The Job class itself can now be used from other java programs.