Change row data structure
Table rows were being saved as Strings, with fields separated by
semicolons (";"). It works well for exporting to csv and to show on
JTables, but for use elsewhere (in external programs, for example) it
can be more precise and more efficient to store data in the original
format, and convert to String when needed.
We created the classes SqliteRow and SqliteElementData to handle the
storage and serialization of this data, and adapted the program to use
it.