Get rid of MappedByteBuffer
FQLite originally reads the whole sqlite file into memory. It is ok to
do that for single files, but for integration in tools that may read
many files that can become a problem.

We tried to get around of this problem by changing the ByteBuffer used
to a MappedByteBuffer (FileChannel.map ...). But memory mapped files
have many problems in Java. For example, it can not be explicitly closed
(one must rely on garbage collector and hope for the file to be closed).
So, we developed another class to encapsulate the reading of a file with
an interface similar to the ByteBuffer used before, in such a way that
it behaves as if the whole file were in memory.

This commit introduces this class (`fqlite.util.RandomAccessFileReader`)
and change the code of fqlite to use it instead of the previous
approach.