Limit amount of allocated byte buffers - FIX OOM
When recovering data from unallocated pages, it is possible to
collect nonsense data. We were trying to allocate buffers with size
grabbed from unallocated pages, relying it is sound. It was causing
negative buffer sizes or huge blocks of memory being allocated.
With this patch we limit the size of the allocated buffer, avoinding
possible OOM errors and NegativeArraySizeException.